+++ native/ios/BulletANE/BulletANE/SphereShapeAPI.cpp
@@ -18,7 +18,7 @@ extern "C" FREObject createSphereShape(FREContext ctx, void *funcData, uint32_t 
     btCollisionShape* shape = new btSphereShape((btScalar)rad);
     
     FREObject ptr;
-    FRENewObjectFromUint32((uint32_t)shape, &ptr);
+    FRENewObjectFromUint32((uint32_t)(size_t)shape, &ptr);
     return ptr;
 }
